Move `autoFailingAsyncTest` into a shared utility accessible to all tests, and create a few other helpers
